+++ great_expectations/data_context/store/__init__.py
@@ -27,4 +27,4 @@ from .query_store import SqlAlchemyQueryStore  # isort:skip
 from .html_site_store import HtmlSiteStore  # isort:skip
 from .profiler_store import ProfilerStore  # isort:skip
 from .datasource_store import DatasourceStore  # isort:skip
-from .data_context_variables_store import DataContextStore  # isort:skip
+from .data_context_store import DataContextStore  # isort:skip
